feat: use ActiveMQ to addOrder

+++ src/main/java/com/yhy/bookstore/daoimpl/OrderDaoImpl.java
@@ -8,29 +8,32 @@ import com.yhy.bookstore.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public class OrderDaoImpl implements OrderDao {
-    @Autowired
-    OrderItemRepository orderItemRepository;
-    @Autowired
-    OrderRepository orderRepository;
+  @Autowired OrderItemRepository orderItemRepository;
+  @Autowired OrderRepository orderRepository;
 
-    @Override
-    public void addOrder(Order order)
-    {
-        orderRepository.saveAndFlush(order);
-    }
+  @Override
+  public void addOrder(Order order) {
+    orderRepository.saveAndFlush(order);
+  }
 
-    @Override
-    public void addOrderItem(OrderItem orderItem)
-    {
-        orderItemRepository.saveAndFlush(orderItem);
-    }
+  @Override
+  public void addOrderItem(OrderItem orderItem) {
+    orderItemRepository.saveAndFlush(orderItem);
+  }
 
-    @Override
-    public void delOrder(Integer orderId)
-    {
-        if(orderRepository.existsById(orderId))
-            orderRepository.deleteById(orderId);
+  @Override
+  public void addOrderItems(List<OrderItem> orderItems) {
+    for (OrderItem item : orderItems) {
+      orderItemRepository.saveAndFlush(item);
     }
+  }
+
+  @Override
+  public void delOrder(Integer orderId) {
+    if (orderRepository.existsById(orderId)) orderRepository.deleteById(orderId);
+  }
 }
